README.md formatting fixes
Some text that was meant to show up as verbatim/code was not showing up
as such due to indentation. Add extra indentation or backticks where
appropriate so that these blocks are properly formatted. For the long
block that contains documentation for directory structure, add backticks
to make it into a code block.