Configure registry_external_url in gitlab.rb
In order to make use of the GitLab internal Docker registry, it needs to
be enabled in gitlab.rb using the registry_external_url option. Set this
option so that the registry is located under the same domain as GitLab
itself but in port 4567. Also expose port 4567 from the Omnibus
container.

This enables us to use the registry for CI/CD pipelines: it will be
possible to build an image, push it to the registry and use that image
for subsequent jobs instead of building the image separately for every
single job. The nice thing about using GitLab's registry specifically is
that it is per project, so each project only has access to their own
images to keep things hygienic.