Skip registry backups, fix backup cleanup
The registry on GitLab is the component that takes up the most space,
but the registry images are actually not that valuable since they are
mostly just built for testing in pipelines. The registry in GitLab
should not be used for sharing images publicly or archiving old images -
it should exist only to serve pipelines. With that in mind, let's not
create backups of the registry images.

Also fix the tmpwatch that runs via crontab to clean up old backups.
Previously we used access time as the basis for cleaning up old backup
files, but this is not so great because updatedb regularly scans files
on the server, as noted in the tmpwatch man page:

"Note that the periodic updatedb file system scans keep the atime of
directories recent."

Change the tmpwatch commands so that they use file modification time
instead (-m).